% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb.R
\name{ext_vb}
\alias{ext_vb}
\title{extended von Bertalanffy growth function}
\usage{
ext_vb(age, ps, setorigin = FALSE)
}
\arguments{
\item{age}{numerical age}

\item{ps}{vectorized parameters in order of c(Linf, K1, K2, tm, t0)}

\item{setorigin}{constrain t_0 to 0 when TRUE}
}
\value{
numerical or vector
}
\description{
extended von Bertalanffy growth function
}
\examples{
\dontrun{
ext_vb(seq(0,5, 0.5), ps = c(300, 0.6, 0.4, 2, 0))
}
}
