% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_vb.R
\name{fit_ext_vb}
\alias{fit_ext_vb}
\title{Fit vb via maximum likelihood method}
\usage{
fit_ext_vb(dat, ps, setorigin = FALSE)
}
\arguments{
\item{dat}{tibble with column `c(Age, Length)`}

\item{ps}{parameters in order of `c(L_inf, K_1, K_2, t_m, t_0)`}

\item{setorigin}{constrain t_0 to 0 when TRUE}
}
\value{
tibble of estimated parameters
}
\description{
Fit vb via maximum likelihood method
}
\examples{
\dontrun{
foo <- tibble::tibble(Age = seq(0, 5, 0.1),
Length = ext_vb(age = seq(0, 5, 0.1),
ps = c(300, 0.6, 0.4, 3, 0)))
fit_ext_vb(foo, ps = c(250, 0.5, 1, 2, 0))
}
}
