% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_vb.R
\name{fit_vb}
\alias{fit_vb}
\title{Fit vb via maximum likelihood method}
\usage{
fit_vb(dat, ps, setorigin = FALSE)
}
\arguments{
\item{dat}{tibble with column `c(Age, Length)`}

\item{ps}{parameters in order of `c(L_inf, K, t_0)`}

\item{setorigin}{constrain t_0 to 0 when TRUE}
}
\value{
tibble of estimated parameters
}
\description{
Fit vb via maximum likelihood method
}
\examples{
\dontrun{
foo <- tibble::tibble(
Age = seq(0, 5, 0.1),
Length = vb(Age, c(300, 0.6, 0)))
fit_vb(dat = foo, ps = c(200, 0.8, 0))
}
}
