% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb.R
\name{vb}
\alias{vb}
\title{von Bertalanffy growth function}
\usage{
vb(age, ps)
}
\arguments{
\item{age}{numerical age}

\item{ps}{vectorized parameters in order of c(Linf, K, t0)}
}
\value{
numerical or vector
}
\description{
von Bertalanffy growth function
}
\examples{
\dontrun{
vb(1, c(300, 0.6, 0))
}
}
